Added wrong tests for the signup and login
