+++ routes/api/users.js
@@ -1,11 +1,10 @@
 import express from 'express';
-import jwt from 'jsonwebtoken';
 import dotenv from 'dotenv';
 
 import { getUser, addUser, loginUser, patchUser, patchAvatar } from '../../models/users.js';
 import { auth } from '../../config/config-passport.js';
 import { uploadImage } from '../../config/config-multer.js';
-import { signupMiddleware } from './users.middleware.js';
+import { signupMiddleware, loginMiddleware } from './users.middleware.js';
 
 dotenv.config();
 const secret = process.env.SECRET;
@@ -32,42 +31,7 @@ usersRouter.get('/current', auth, async (req, res, next) => {
 
 usersRouter.post('/signup', signupMiddleware(addUser));
 
-usersRouter.post('/login', async (req, res, next) => {
-  const { body } = req;
-
-  if (!('email' in body) || !('password' in body)) {
-    return res.status(400).json('Error! Missing password or email field!');
-  }
-
-  try {
-    const user = await loginUser(body);
-
-    if (!user) {
-      return res.status(400).json(`Error! Email or password is wrong!`);
-    }
-
-    const payload = {
-      id: user.id,
-      username: user.email,
-    };
-
-    const token = jwt.sign(payload, secret, { expiresIn: '1h' });
-
-    user.token = token;
-    await user.save();
-
-    const { email, subscription } = user;
-
-    res.status(200).json({
-      status: 'success',
-      code: 200,
-      token: token,
-      user: { email, subscription },
-    });
-  } catch (err) {
-    res.status(500).json(`An error occurred while logging the user! ${err}`);
-  }
-});
+usersRouter.post('/login', loginMiddleware(loginUser));
 
 usersRouter.post('/logout', auth, async (req, res, next) => {
   try {
